Visit nearest point first
This reduces "trashing" of the binary heap, as it visits points that are more likely to be nearests first, and then compares the rest with those found. Previously it would fill up binary heap quickly along the way, and then start swapping items, resulting in a lot more swaps than necessary.

On some real-world examples, this showed ~20% increase to kNN search.